refactored a litlle more
